Rename template-reducer
