--- src/redux/template-reducer/template-reducer.js
@@ -1,38 +1,0 @@
-import SomeService from '../../services/SomeService';
-
-const SOME_ACTION = 'template/template-reducer/SOME_ACTION';
-const ANOTHER_ACTION = 'template/template-reducer/ANOTHER_ACTION';
-
-async function doSomething(dispatch, getState) {
-  const { stuff: currentStuffs } = getState();
-
-  if (currentStuffs.length === 0) {
-    const { data } = await SomeService.getStuffs();
-    const fetchedStuffs = data.map((stuff) => ({
-      fetchedValue: stuff.value,
-      default: 'default value',
-    }));
-
-    dispatch({ type: SOME_ACTION, payload: fetchedStuffs });
-  }
-}
-
-function doAnotherThing(data) {
-  return {
-    type: ANOTHER_ACTION,
-    payload: data,
-  };
-}
-
-export default function reducer(state = [], action) {
-  switch (action.type) {
-    case SOME_ACTION:
-      return state;
-    case ANOTHER_ACTION:
-      return state;
-    default:
-      return state;
-  }
-}
-
-export { doSomething, doAnotherThing };
